Added pathfinding
Pathfinding is done using navigation meshes.
Estimated performance is about 20ms for long paths, and minimal for short paths.
In some cases the algorithm does not choose the absolutely shortest path,
but the difference should be negligible.

Positions of other units on the map are NOT taken into account during pathfinding.