Found and fixed some bugs in Recall
Previously you could Recall that created a timer, and if you stopped halfway moved a bit and started recalling again you would actually recall. So I added a new "wasInterrupted" variable that checks if the Recall was interrupted so you won't be able to replicate the bug. Also I removed the posibillity to spam the recall if you stayed in the same place.  Also I was getting an error about the UpdateBuff method in BuffGameScriptController ( used in MissFortuneStruct ). I realized that it lacked such a method and added an empty one for the script to load and work but this has to be looked into.