Added new events for ApiEventManager, support for Grievous Wounds, and a class for Utility Helper functions (#6)
With the addition of OnAutoAttackHit, abilities that work with on-hit effects such as Miss Fortune's Impure Shots should now be properly handled. Additionally, support for Grievous Wounds means that abilities that reduce healing effects ,may now function fully. Finally, Helpers.cs will be used for various utility functions, such as Rotate, that are useful for abilities such as Twisted Fate and Graves's Q abilities, which fire in cones.