Fix the rest of vision (mostly)
Fixed sometimes losing vision of a unit when the raycast stops where it started.
Fixed not being able to see through the NavGridFlag 0x40, this consequentially fixes not being able to attack units that are behind a see through part of the NavGrid, such as Inhibitors and Nexus'.
Fixed being able to see through bushes when you are not in one as well as fixed being able to see through bushes that are not in the direction of a unit getting a ray casted to it (while you are in a bush).