DebugMode using Particles & Comments
Adds the !debugmode command which has the option to draw the player's server-side collision radius and waypoints. This can be expanded to other units such as projectiles and minions.
Streamlined the process of adding, removing, and networking particles.
Added more variables to particles related to networking.
Cleaned the GameObject.Move function and added some functions related to movement.
Cleaned the NotifyFXCreateGroup function and used new Particle variables.
Added comments to GameObject and Particle.
Shifted location of adding players in Game so that PacketNotifier never causes an exception.