Resolve #1306, Progress #1326 -- Vision and spawn system rework (#1329)
Previously, the vision system tried to be team-based. The proposed one still calculates team vision, but sends out packets based on players' individual vision, which is linked to their champions' vision. This will make it possible to implement such effects as nearsight in the future.
- It is now possible to describe all the conditions under which one unit sees another in one place, in the `UnitHasVisionOn` function of the `ObjectManager`.
- Some objects now only spawn when entering player line of sight to prevent cheating:
`spawnShouldBeHidden = obj is IParticle || obj is ISpellMissile || (obj is IMinion && !(obj is ILaneMinion))`
- Part of the functionality has been moved from `HandleStartGame` to `HandleSpawn`. Now it is possible to reconnect and use skills and reconnect during a pause (still have bugs with minimap, inventory and time)